Fix TS errors
