+++ src/handlers/platforms/next/process-protected-handler.ts
@@ -3,6 +3,7 @@ import { NextApiRequest } from "next";
 
 import { APL } from "@/APL";
 import { createDebug } from "@/debug";
+import { SaleorProtectedHandlerError } from "@/handlers/shared/protected-handler";
 import { ProtectedHandlerContext } from "@/handlers/shared/protected-handler-context";
 import { getBaseUrl, getSaleorHeaders } from "@/headers";
 import { getOtelTracer } from "@/open-telemetry";
@@ -11,16 +12,6 @@ import { extractUserFromJwt } from "@/util/extract-user-from-jwt";
 import { verifyJWT } from "@/verify-jwt";
 
 const debug = createDebug("processProtectedHandler");
-
-export type SaleorProtectedHandlerError =
-  | "OTHER"
-  | "MISSING_HOST_HEADER"
-  | "MISSING_DOMAIN_HEADER"
-  | "MISSING_API_URL_HEADER"
-  | "MISSING_AUTHORIZATION_BEARER_HEADER"
-  | "NOT_REGISTERED"
-  | "JWT_VERIFICATION_FAILED"
-  | "NO_APP_ID";
 
 export class ProtectedHandlerError extends Error {
   errorType: SaleorProtectedHandlerError = "OTHER";
