+++ src/handlers/shared/validate-allow-saleor-urls.ts
@@ -1,4 +1,4 @@
-import { CreateAppRegisterHandlerOptions } from "../next/create-app-register-handler";
+import { CreateAppRegisterHandlerOptions } from "../platforms/next/create-app-register-handler";
 
 export const validateAllowSaleorUrls = (
   saleorApiUrl: string,
